Update State-Declaration of Items
mainpage.py now uses an Enum-Class for the Item States
instead of simple Strings like "normal", "edit" and "create".

This changed the Item Behaviour on State CREATE for some reason.
The Buttons didn't instantly appear on newly created Items as they should.
Therefore, the addition of CREATE-Checking in change_state()
to place the Buttons.